Don't set TPU topology env vars in cloud_tpu_init.py
This used to be necessary. However, now these are automatically set in
libtpu. Beyond being redundant, the Python logic needs to be updated
to avoid getting KeyErrors on new topologies and TPU versions, so
better to remove it.

This also moves `get_metadata` to cloud_tpu_cluster.py since it's only
used in that file now.
